requirements.txt
fastapi>=0.115.6
langchain>=0.3.14
langchain-community>=0.3.14
langchain-openai>=0.3.1
langchain-chroma>=0.2.0
unstructured>=0.16.13
python-dotenv>=1.0.1
loguru>=0.7.3
chromadb>=0.6.3
python-docx>=1.1.2
rank-bm25>=0.2.2
python-multipart>=0.0.20
redis>=5.2.1
python-magic-bin>=0.4.14
markdown>=3.7
sentence-transformers>=2.2.2
torch>=2.1.0

